+++ bash/incl/port4me.bash
@@ -1,45 +1,81 @@
 #! /usr/bin/env bash
 
-port4me_user() {
-    local res=${PORT4ME_USER}
-    [[ -z $res ]] && res=${USER}
-    echo "$res"
+declare -i LCG_SEED
+export LCG_SEED
+
+p4m_error() {
+    >&2 echo "ERROR: $1"
+    exit 1
 }
 
-port4me_tool() {
-    local res=${PORT4ME_TOOL}
-    echo "$res"
-}
+#' Check if TCP port can be opened
+#'
+#' Examples:
+#' can_port_be_opened 4001
+#' openable=$?
+#'
+#' Requirements:
+#' * either 'nc' or 'ss'
+PORT4ME_PORT_COMMAND=
+p4m_can_port_be_opened() {
+    local -i port=${1:?}
+    local cmds=(nc ss)
+    local cmd
+    
+    (( port < 1 || port > 65535 )) && p4m_error "Port is out of range [1,65535]: ${port}"
 
-port4me_seed() {
-    local user tool seed_str
-    local -i seed
-    
-    user=$(port4me_user)
-    tool=$(port4me_tool)
-    
-    seed_str=
-    if [[ -n $user &&  -n $tool ]]; then
-        seed_str="$user,$tool"
-    elif [[ -n $user ]]; then
-        seed_str="$user"
-    elif [[ -n $tool ]]; then
-        seed_str="$tool"
-    else
-        error "At least one of arguments 'user' and 'tool' must be non-empty"
+    ## Identify port command and memoize, unless already done
+    if [[ -z ${PORT4ME_PORT_COMMAND} ]]; then
+        for cmd in "${cmds[@]}"; do
+            if command -v "${cmd}" > /dev/null; then
+                PORT4ME_PORT_COMMAND=${cmd}
+                break
+            fi
+        done
+        [[ -z ${PORT4ME_PORT_COMMAND} ]] && p4m_error "Cannot check if port is available or not. None of the following commands exist on this system: ${cmds[*]}"
     fi
-
-    seed=$(string_to_uint "$seed_str")
-
-    if ${PORT4ME_DEBUG:-false}; then
-       >&2 printf "seed_str='%s'\n" "$seed_str"
-       >&2 printf "seed=%d\n" "$seed"
+    
+    ## Is port occupied?
+    if [[ ${PORT4ME_PORT_COMMAND} == "nc" ]]; then
+        if nc -z 127.0.0.1 "$port"; then
+            return 1
+        fi
+    elif [[ ${PORT4ME_PORT_COMMAND} == "ss" ]]; then
+        if ss -H -l -n src :"$port" | grep -q -E ":$port\b"; then
+            return 1
+        fi
     fi
     
-    echo "$seed"
+    ## FIXME: A port can be free, but it might be that the user
+    ## don't have the right to open it, e.g. port 1-1023.
+    ## WORKAROUND: If non-root, assume 1-1023 can't be opened
+    if [[ "$EUID" != 0 ]]; then
+        if (( port < 1024 )); then
+            return 1
+        fi
+    fi
+    
+    return 0
 }
 
-parse_ports() {
+#' Analogue to Java hashCode() but returns a non-signed integer
+p4m_string_to_uint() {
+    local str="$1"
+    local -i kk byte
+    local -i hash=0
+    
+    for ((kk = 0; kk < ${#str}; kk++)); do
+        ## ASCII character to ASCII value
+        LC_TYPE=C printf -v byte "%d" "'${str:$kk:1}"
+        hash=$(( 31 * hash + byte ))
+        ## Corce to non-signed integer [0,2^32-1]
+        hash=$(( hash % 2**32 ))
+    done
+    
+    printf "%d" $hash
+}
+
+p4m_parse_ports() {
     local spec=${1:?}
     local specs
     local -a ports
@@ -71,47 +107,71 @@ parse_ports() {
     fi
 }
 
-port4me_prepend() {
-    parse_ports "${PORT4ME_PREPEND},${PORT4ME_PREPEND_SITE}"
-}    
+p4m_lcg() {
+    local -i a=75 c=74 modulus=65537 seed="${LCG_SEED:?}"
+    local -i seed_next
 
-port4me_include() {
-    parse_ports "${PORT4ME_INCLUDE},${PORT4ME_INCLUDE_SITE}"
-}    
+    ## Make sure seed is within [0,modulus-1] to avoid integer overflow
+    seed=$(( seed % modulus ))
 
-port4me_exclude() {
-    parse_ports "${PORT4ME_EXCLUDE},${PORT4ME_EXCLUDE_SITE}"
-}    
+    seed_next=$(( (a * seed + c) % modulus ))
+
+    ## For certain LCG parameter settings, we might end up in the same
+    ## LCG state. For example, this can happen when (a-c) = 1 and
+    ## seed = modulus-1. To make sure we handle any parameter setup, we
+    ## detect this manually, increment the seed, and recalculate.
+    if (( seed_next == seed )); then
+        seed=$(( seed + 1 ))
+        seed_next=$(( (a * seed + c) % modulus ))
+    fi
+
+    ## Sanity checks
+    if (( seed_next < 0 )); then
+        p4m_error "INTERNAL: New LCG seed is non-positive: $seed_next, where (a, c, modulus) = ($a, $c, $modulus) with seed = $seed"
+    elif (( seed_next > modulus )); then
+        p4m_error "INTERNAL: New LCG seed is too large: $seed_next, where (a, c, modulus) = ($a, $c, $modulus) with seed = $seed"
+    elif (( seed_next == seed )); then
+        p4m_error "INTERNAL: New LCG seed is same a current seed, where (a, c, modulus) = ($a, $c, $modulus) with seed = $seed"
+    fi
+    
+    LCG_SEED=${seed_next}
+    
+    echo "${LCG_SEED}"
+}
+
+p4m_string_to_seed() {
+    local seed=${PORT4ME_USER:-${USER:?}},${PORT4ME_TOOL}
+    seed=${seed%%,}  ## trim trailing commas
+    p4m_string_to_uint "$seed"
+}
 
 port4me() {
-    local -i skip=${PORT4ME_SKIP:-0}
-    local -i exclude
-    local -i include
-    local -i prepend
-    local -i count
-    local -i list=${PORT4ME_LIST:-0}
-    local -i max_tries=${PORT4ME_MAX_TRIES:-1000}
+    local -i max_tries=${PORT4ME_MAX_TRIES:-65535}
     local must_work=${PORT4ME_MUST_WORK:-true}
+    local -i skip=${PORT4ME_SKIP:-0}
+    local -i list=${PORT4ME_LIST:-0}
+    local -i exclude include prepend
+    local -i count tries
 
-    mapfile -t exclude < <(port4me_exclude)
-    mapfile -t include < <(port4me_include)
-    mapfile -t prepend < <(port4me_prepend)
+    mapfile -t exclude < <(p4m_parse_ports "${PORT4ME_EXCLUDE},${PORT4ME_EXCLUDE_SITE}")
+    mapfile -t include < <(p4m_parse_ports "${PORT4ME_INCLUDE},${PORT4ME_INCLUDE_SITE}")
+    mapfile -t prepend < <(p4m_parse_ports "${PORT4ME_PREPEND},${PORT4ME_PREPEND_SITE}")
 
     if (( list > 0 )); then
         max_tries=${list}
     fi
     
-    lcg_set_params
-    lcg_set_seed "$(port4me_seed)"
+    LCG_SEED=$(p4m_string_to_seed)
 
     count=0
-    while (( count < max_tries )); do
+    tries=0
+    while (( tries < max_tries )); do
         if (( ${#prepend[@]} > 0 )); then
             port=${prepend[0]}
-            (( port < 1 || port > 65535 )) && error "Prepended port out of range [1,65535]: ${port}"
+            (( port < 1 || port > 65535 )) && p4m_error "Prepended port out of range [1,65535]: ${port}"
             prepend=("${prepend[@]:1}") ## drop first element
         else
-            lcg_port > /dev/null
+            p4m_lcg > /dev/null
             port=${LCG_SEED:?}
         fi
 
@@ -129,8 +189,11 @@ port4me() {
                 ${PORT4ME_DEBUG:-false} && >&2 printf "Port not included: %d\n" "$port"
                 continue
             fi
+        elif (( LCG_SEED < 1024 || LCG_SEED > 65535 )); then
+            continue
         fi
 
+        tries=$(( tries + 1 ))
         count=$((count + 1))
 
         if (( list > 0 )); then
@@ -143,7 +206,7 @@ port4me() {
             
             ${PORT4ME_DEBUG:-false} && >&2 printf "%d. port=%d\n" "$count" "$port"
     
-            if can_port_be_opened "$port"; then
+            if p4m_can_port_be_opened "$port"; then
                 printf "%d\n" "$port"
                 return 0
             fi
@@ -153,7 +216,7 @@ port4me() {
 
     if (( list == 0 )); then
         if $must_work; then
-            error "Failed to find a free TCP port"
+            p4m_error "Failed to find a free TCP port after ${max_tries} attempts"
         fi
 
         printf "%d\n" "-1"