+++ r/R/port4me.R
@@ -53,8 +53,7 @@ port4me_prepend <- function() {
 
   for (name in c("PORT4ME_PREPEND", "PORT4ME_PREPEND_SITE")) {
     arg <- Sys.getenv(name, "")
-    ports <- parse_ports(arg)
-    ports <- c(ports, ports)
+    ports <- c(ports, parse_ports(arg))
   }
   ports <- unique(ports)
 
@@ -66,8 +65,7 @@ port4me_exclude <- function() {
 
   for (name in c("PORT4ME_EXCLUDE", "PORT4ME_EXCLUDE_SITE")) {
     arg <- Sys.getenv(name, "")
-    ports <- parse_ports(arg)
-    ports <- c(ports, ports)
+    ports <- c(ports, parse_ports(arg))
   }
   ports <- unique(ports)
   
@@ -79,8 +77,7 @@ port4me_include <- function() {
 
   for (name in c("PORT4ME_INCLUDE", "PORT4ME_INCLUDE_SITE")) {
     arg <- Sys.getenv(name, "")
-    ports <- parse_ports(arg)
-    ports <- c(ports, ports)
+    ports <- c(ports, parse_ports(arg))
   }
   ports <- unique(ports)
   
@@ -93,7 +90,41 @@ port4me_skip <- function() {
   skip
 }
 
-port4me <- function(user = port4me_user(), tool = port4me_tool(), prepend = port4me_prepend(), include = port4me_exclude(), exclude = port4me_exclude(), skip = port4me_skip(), list = NULL, test = NULL, max_tries = 1000L, must_work = TRUE) {
+
+#' Gets a personalized TCP port that can be opened
+#'
+#' @param user (optional) The name of the user.
+#' Defaults to `Sys.info()[["user"]]`.
+#'
+#' @param tool (optional) The name of the software tool for which a port
+#' should be generated.
+#'
+#' @param prepend (optional) An integer vector of ports to always consider.
+#'
+#' @param include (optional) An integer vector of possible ports to return.
+#' Defaults to `1024::65535`.
+#'
+#' @param exclude (optional) An integer vector of ports to exclude.
+#'
+#' @param skip (optional) Number of non-excluded ports to skip.
+#'
+#' @param list (optional) Number of ports to list.
+#'
+#' @param test (optional) A port to check whether it can be opened or not.
+#'
+#' @param max_tries Maximum number of ports checked, before giving up.
+#'
+#' @param must_work If TRUE, then an error is produced if no port could
+#' be found.  If FALSE, then `-1` is returned.
+#'
+#' @return
+#' An port or a vector of ports.
+#' If `test` is given, then TRUE is if the port can be opened, otherwise FALSE.
+#'
+#' @example incl/port4me.R
+#'
+#' @export
+port4me <- function(user = port4me_user(), tool = port4me_tool(), prepend = port4me_prepend(), include = port4me_include(), exclude = port4me_exclude(), skip = port4me_skip(), list = NULL, test = NULL, max_tries = 65535L, must_work = TRUE) {
   stopifnot(length(user) == 1L, is.character(user), !is.na(user))
   stopifnot(is.null(tool) || is.character(tool), !anyNA(tool))
   if (!is.null(list)) {
@@ -124,9 +155,18 @@ port4me <- function(user = port4me_user(), tool = port4me_tool(), prepend = port
 
   if (!is.null(list)) max_tries <- list + skip
 
+  if (isTRUE(as.logical(Sys.getenv("PORT4ME_DEBUG", "false")))) {
+    utils::str(list(
+      include = include,
+      exclude = exclude,
+      prepend = prepend
+    ))
+  }
+
   ports <- integer(0)
   count <- 0L
-  while (count <= max_tries) {
+  tries <- 0L
+  while (tries <= max_tries) {
     if (length(prepend) > 0) {
       port <- prepend[1]
       prepend <- prepend[-1]
@@ -135,6 +175,7 @@ port4me <- function(user = port4me_user(), tool = port4me_tool(), prepend = port
     }
     if (port %in% exclude) next
     if (length(include) > 0 && (! port %in% include)) next
+    tries <- tries + 1L
     count <- count + 1L
     if (count <= skip) next
     if (is.null(list)) {
@@ -145,7 +186,9 @@ port4me <- function(user = port4me_user(), tool = port4me_tool(), prepend = port
     }
   }
 
-  if (must_work) stop("Failed to find a free TCP port")
+  if (must_work) {
+    stop(sprintf("Failed to find a free TCP port after %d attempts", max_tries))
+  }
 
   -1L
 }